MWPW-125675 - [Core] Milo Bulk Publish
- split logic and presentation code
