+++ libs/blocks/bulk/bulk.js
@@ -1,242 +1,24 @@
 import { html, render, useState, useRef } from '../../deps/htm-preact.js';
-import { getImsToken } from '../../../tools/send-to-caas/send-utils.js';
-import { fetchWithTimeout } from '../caas/utils.js';
-import { loadScript } from '../../utils/utils.js';
-import {
-  getLocalStorage,
-  setLocalStorage,
-} from '../review/utils/localStorageUtils.js';
 import createCopy from '../library-config/library-utils.js';
+import {
+  signOut,
+  getStoredUrlInput,
+  getActionName,
+  getUrls,
+  userIsAuthorized,
+  executeActions,
+  getCompletion,
+  sendReport,
+  signIn,
+  getUser,
+  getStoredOperation,
+  storeUrls,
+  storeOperation,
+} from './utils.js';
 
 // TODO enable IMS sign in?
 const IMS_SIGN_IN_ENABLED = false;
-const ADMIN_BASE_URL = 'https://admin.hlx.page';
 const URLS_ENTRY_LIMIT = 1000;
-const THROTTLING_DELAY_MS = 100;
-const BULK_CONFIG_FILE_PATH = '/tools/bulk-config.json';
-const BULK_REPORT_FILE_PATH = '/tools/bulk-report';
-const BULK_AUTHORIZED_USERS = 'bulkAuthorizedUsers';
-const BULK_SUPPORTED_SITES = 'bulkSupportedSites';
-const BULK_STORED_URLS = 'bulkStoredUrls';
-const BULK_STORED_URL_IDX = 'bulkStoredUrlIdx';
-const BULK_STORED_COMPLETED = 'bulkStoredCompleted';
-const BULK_STORED_OPERATION_NAME = 'bulkStoredOperationName';
-const UNSUPPORTED_SITE = 'unsupported domain';
-
-const getUser = async () => {
-  const profile = await window.adobeIMS?.getProfile();
-  return profile
-    ? { name: profile.name, email: profile.email }
-    : { name: 'anonymous', email: 'anonymous' };
-};
-
-const signIn = async () => {
-  if (!IMS_SIGN_IN_ENABLED) return true;
-  // force user to sign in
-  const accessToken = await getImsToken(loadScript);
-  if (!accessToken) {
-    window.adobeIMS.signIn();
-    return false;
-  }
-  return true;
-};
-
-const getAuthorizedUsers = async () => {
-  let authorizedUsers = getLocalStorage(BULK_AUTHORIZED_USERS);
-  if (authorizedUsers) return authorizedUsers;
-  const resp = await fetch(BULK_CONFIG_FILE_PATH);
-  const json = await resp.json();
-  authorizedUsers = json.users.data.map((user) => user.user);
-  setLocalStorage(BULK_AUTHORIZED_USERS, authorizedUsers);
-  return authorizedUsers;
-};
-
-const getSupportedSites = async () => {
-  let supportedSites = getLocalStorage(BULK_SUPPORTED_SITES);
-  if (supportedSites) return supportedSites;
-  const resp = await fetch(BULK_CONFIG_FILE_PATH);
-  const json = await resp.json();
-  supportedSites = json.sites.data.map((site) => site.origin);
-  setLocalStorage(BULK_SUPPORTED_SITES, supportedSites);
-  return supportedSites;
-};
-
-const userIsAuthorized = async () => {
-  const { email } = await getUser();
-  const users = await getAuthorizedUsers();
-  return users.includes(email);
-};
-
-const siteIsSupported = async (url) => {
-  try {
-    const { origin } = new URL(url);
-    const sites = await getSupportedSites();
-    return sites.includes(origin);
-  } catch (e) {
-    return false;
-  }
-};
-
-const getUrls = (element) => element.current?.value.split('\n').filter((url) => url.length > 0).map((e) => e.trim());
-
-const getActionName = (action, gerund) => {
-  let name;
-  switch (action) {
-    case null:
-    case 'preview':
-      name = (!gerund) ? 'Preview' : 'Previewing';
-      break;
-    case 'publish':
-      name = (!gerund) ? 'Publish' : 'Publishing';
-      break;
-    default:
-      name = (!gerund) ? 'Preview & publish' : 'Previewing & publishing';
-  }
-  return name;
-};
-
-const executeAction = async (action, url) => {
-  const operation = (action === 'preview') ? 'preview' : 'live';
-  const siteAllowed = await siteIsSupported(url);
-  if (!siteAllowed) return UNSUPPORTED_SITE;
-  const { hostname, pathname } = new URL(url);
-  const [branch, repo, owner] = hostname.split('.')[0].split('--');
-  const adminURL = `${ADMIN_BASE_URL}/${operation}/${owner}/${repo}/${branch}${pathname}`;
-  const resp = await fetchWithTimeout(adminURL, { method: 'POST' });
-  return resp.status;
-};
-
-// eslint-disable-next-line no-promise-executor-return
-const delay = (milliseconds) => new Promise((resolve) => setTimeout(resolve, milliseconds));
-
-const executeActions = async (actions, urls, setResult, lastUrlIdx = 0) => {
-  setLocalStorage(BULK_STORED_COMPLETED, false);
-  const results = [];
-  if (lastUrlIdx > urls.length - 1) {
-    // eslint-disable-next-line no-console
-    console.error(`incorrect url index: ${lastUrlIdx}`);
-    return null;
-  }
-  // eslint-disable-next-line no-plusplus
-  for (let i = lastUrlIdx; i < urls.length; i++) {
-    const url = urls[i];
-    const status = {};
-    // eslint-disable-next-line no-plusplus
-    for (let j = 0; j < actions.length; j++) {
-      const action = actions[j];
-      // eslint-disable-next-line no-await-in-loop
-      status[action] = await executeAction(action, url);
-      // eslint-disable-next-line no-await-in-loop
-      await delay(THROTTLING_DELAY_MS);
-    }
-    results.push({
-      url,
-      status,
-    });
-    setResult([...results]);
-    setLocalStorage(BULK_STORED_URL_IDX, i);
-  }
-  setLocalStorage(BULK_STORED_COMPLETED, true);
-  return results;
-};
-
-const getCompletion = (results) => {
-  let previewTotal = 0;
-  let publishTotal = 0;
-  let previewSuccess = 0;
-  let publishSuccess = 0;
-  results.forEach((result) => {
-    const { status } = result;
-    if (status.preview) previewTotal += 1;
-    if (status.publish) publishTotal += 1;
-    if (status.preview === 200) previewSuccess += 1;
-    if (status.publish === 200) publishSuccess += 1;
-  });
-  return {
-    preview: {
-      total: previewTotal,
-      success: previewSuccess,
-    },
-    publish: {
-      total: publishTotal,
-      success: publishSuccess,
-    },
-  };
-};
-
-const getReport = async (results, action) => {
-  const origins = {};
-  results.forEach((result) => {
-    let origin;
-    try {
-      const urlObj = new URL(result.url);
-      origin = urlObj.origin;
-    } catch (e) {
-      origin = result.url;
-    }
-    if (!origins[origin]) {
-      origins[origin] = {
-        total: 0,
-        success: 0,
-      };
-    }
-    if (action === 'preview&publish') {
-      origins[origin].total += 2;
-    } else {
-      origins[origin].total += 1;
-    }
-    if (result.status.preview === 200) {
-      origins[origin].success += 1;
-    }
-    if (result.status.publish === 200) {
-      origins[origin].success += 1;
-    }
-  });
-  const timestamp = new Date().toISOString();
-  const { email } = await getUser();
-  return Object.keys(origins).map((origin) => ({
-    timestamp,
-    email,
-    action: getActionName(action),
-    domain: origin,
-    urls: origins[origin].total,
-    success: origins[origin].success,
-  }));
-};
-
-const sendReport = async (results, action) => {
-  const rows = await getReport(results, action);
-  rows.forEach((row) => {
-    fetch(BULK_REPORT_FILE_PATH, {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ data: row }),
-    });
-  });
-};
-
-const getStoredOperation = () => {
-  const storedOperationName = getLocalStorage(BULK_STORED_OPERATION_NAME);
-  const storedUrlIdx = getLocalStorage(BULK_STORED_URL_IDX);
-  const storedUrls = getLocalStorage(BULK_STORED_URLS);
-  const completed = getLocalStorage(BULK_STORED_COMPLETED);
-  return {
-    storedOperationName,
-    storedUrlIdx,
-    storedUrls,
-    completed,
-  };
-};
-
-const getStoredUrlInput = () => {
-  const storedUrls = getLocalStorage(BULK_STORED_URLS);
-  return storedUrls ? storedUrls.join('\n') : '';
-};
-
-const signOut = (e) => {
-  e.preventDefault();
-  window.adobeIMS?.signOut();
-};
 
 function User({ user }) {
   return html`
@@ -256,7 +38,7 @@ function UrlInput({ urlsElt }) {
   `;
 }
 
-function SelectBtn({ actionElt, onSelectChange, storedOperationName}) {
+function SelectBtn({ actionElt, onSelectChange, storedOperationName }) {
   return html`
       <select class="bulk-action-select" ref="${actionElt}" onChange=${onSelectChange}>
           <option value="preview">Preview</option>
@@ -431,8 +213,8 @@ function Bulk({ user, storedOperation }) {
   const [authorized, setAuthorized] = useState(true);
   const [urlNumber, setUrlNumber] = useState(-1);
   const [bulkTriggered, setBulkTriggered] = useState(false);
-  const [selectedAction, setSelectedAction] = useState(storedOperation.storedOperationName);
-  const [submittedAction, setSubmittedAction] = useState(storedOperation.storedOperationName);
+  const [selectedAction, setSelectedAction] = useState(storedOperation.name);
+  const [submittedAction, setSubmittedAction] = useState(storedOperation.name);
   const [result, setResult] = useState(null);
   const [completion, setCompletion] = useState(null);
 
@@ -445,7 +227,7 @@ function Bulk({ user, storedOperation }) {
 
   const displayResumeDialog = (storedOperation.completed === null || storedOperation.completed) ? '' : 'displayed';
 
-  const executeBulk = async (operation, startUrlIdx) => {
+  const executeBulk = async (operation, startUrlIdx = 0) => {
     // reset the result area
     setResult(null);
     setCompletion(null);
@@ -460,7 +242,7 @@ function Bulk({ user, storedOperation }) {
 
     // validate the number of urls
     const urls = getUrls(urlsElt);
-    setLocalStorage(BULK_STORED_URLS, urls);
+    storeUrls(urls);
     const urlNumberValue = urls.length - startUrlIdx;
     setUrlNumber(urlNumberValue);
     if (urlNumberValue < 1 || urlNumberValue > URLS_ENTRY_LIMIT) {
@@ -500,16 +282,15 @@ function Bulk({ user, storedOperation }) {
     setBulkTriggered(true);
     const operation = actionElt.current.value;
     setSubmittedAction(operation);
-    setLocalStorage(BULK_STORED_OPERATION_NAME, operation);
-    await executeBulk(operation, 0);
+    storeOperation(operation);
+    await executeBulk(operation);
   };
 
   const resume = async () => {
     setBulkTriggered(true);
     hideModal();
-    const operation = storedOperation.storedOperationName;
-    const urlIdx = storedOperation.storedUrlIdx;
-    await executeBulk(operation, urlIdx + 1);
+    const startUrlIdx = storedOperation.urlIdx !== null ? storedOperation.urlIdx + 1 : 0;
+    await executeBulk(storedOperation.name, startUrlIdx);
   };
 
   return html`
@@ -529,7 +310,7 @@ function Bulk({ user, storedOperation }) {
                 <${SelectBtn}
                     actionElt=${actionElt}
                     onSelectChange=${onSelectChange}
-                    storedOperationName=${storedOperation.storedOperationName} />
+                    storedOperationName=${storedOperation.name} />
             </div>
         </div>
         <${ErrorMessage}
@@ -557,8 +338,10 @@ function Bulk({ user, storedOperation }) {
 }
 
 export default async function init(el) {
-  const signedIn = await signIn();
-  if (!signedIn) return;
+  if (IMS_SIGN_IN_ENABLED) {
+    const signedIn = await signIn(IMS_SIGN_IN_ENABLED);
+    if (!signedIn) return;
+  }
 
   const user = await getUser();
   const storedOperation = getStoredOperation();
